requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
asgiref==3.8.1
    # via django
django==5.2
    # via
    #   sistema-estacionamento (pyproject.toml)
    #   django-jazzmin
    #   djangorestframework
    #   djangorestframework-simplejwt
django-jazzmin==3.0.1
    # via sistema-estacionamento (pyproject.toml)
djangorestframework==3.16.0
    # via
    #   sistema-estacionamento (pyproject.toml)
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.0
    # via sistema-estacionamento (pyproject.toml)
dotenv==0.9.9
    # via sistema-estacionamento (pyproject.toml)
pyjwt==2.9.0
    # via djangorestframework-simplejwt
python-dotenv==1.1.0
    # via dotenv
sqlparse==0.5.3
    # via django
